refactor: changed files structure for the player form
