+++ src/app/register/form/page.tsx
@@ -1,5 +1,5 @@
-import { FormSetpOne } from '@/features/register/FormStepOne'
-import { FormLayout } from '@/features/register/FormLayout'
+import { FormSetpOne } from '@/features/player-registration/FormStep1/FormStepOne'
+import { FormLayout } from '@/features/player-registration/FormStep1/FormLayout'
 
 export default function Form() {
   return (
