+++ src/features/player-registration/FormStep1/FormLayout.tsx
