+++ src/features/player-registration/FormStep1/components/FormTitle.tsx
