+++ src/features/player-registration/FormStep1/components/InputIcon.tsx
