+++ src/features/player-registration/FormStep1/helper/badWords.ts
