+++ src/features/player-registration/FormStep1/helper/badWordsFilter.ts
