+++ src/features/player-registration/FormStep1/icons/Arrow.tsx
