+++ src/features/player-registration/FormStep1/icons/Done.tsx
