+++ src/features/player-registration/FormStep1/icons/NotValid.tsx
