+++ src/features/player-registration/FormStep1/icons/Update.tsx
