+++ src/features/player-registration/FormStep1/validation.ts
